Socket read timeout value now passed to SocketFactory, read timeout type changed to int from long.

+++ svnkit/src/org/tmatesoft/svn/core/auth/ISVNAuthenticationManager.java
@@ -263,7 +263,7 @@ public interface ISVNAuthenticationManager {
      * @return             connection timeout value
      * @since   1.2
      */
-    public long getReadTimeout(SVNRepository repository);
+    public int getReadTimeout(SVNRepository repository);
     
     public int getConnectTimeout(SVNRepository repository);
 }
